requirements.txt
# Core dependencies
influxdb==5.3.1
Adafruit_DHT==1.4.0
RPi.GPIO==0.7.1

# Development dependencies (optional)
pytest==7.3.1
pytest-cov==4.1.0
black==23.3.0
isort==5.12.0
mypy==1.3.0 